FEMClient

Module description

StatFEM_EUCLID.FEMClient — Module

FEMClient

This submodule contains wrapper functions around UMBridge.jl` for conversion between Julia types and what UMBridge.jl expects or returns.

Functions

StatFEM_EUCLID.FEMClient.evaluate_fem_model — Method

evaluate_fem_model(fem_model, parameter::Float64, solution_index = 1)

Query the server to solve the FEM model fem_model for a given (material) parameter. Keyword arguments:

solution_index: For a vector valued unknown, return the unknown at this index (can be : for the full solution)

config: Dict{String,Any} describing optional parameters for the fem_model

StatFEM_EUCLID.FEMClient.evaluate_fem_model — Method

evaluate_fem_model(fem_model::HTTPModel, parameters::Vector{Float64}; solution_index = 1, config::Dict{String,Any} = empty_config())

Query the server to solve the FEM model fem_model for a given set of (material) parameters. Keyword arguments:

solution_index: For a vector valued unknown, return the unknown at this index (can be : for the full solution)

config: Dict{String,Any} describing optional parameters for the fem_model