PCE

StatFEM_EUCLID.PCE — Module

PCE

This submodule provides all the functionality for building a PCE surrogate from an FEM sample obtained through the Sampling submodule.

Obtaining the PCE surrogate

StatFEM_EUCLID.PCE.PolyChaosExpansion — Type

PolyChaosExpansion{T<: Real}

Struct for holding the PCE surrogate

StatFEM_EUCLID.PCE.PolyChaosExpansion — Method

PolyChaosExpansion(sample::UnivariateFEMSample{T};polynomials::AbstractCanonicalOrthoPoly = GaussOrthoPoly(8),vector_distribution::UnivariateDistribution=Normal(0,1)) where T<:Real

Represent the sampled random variable $Y$(sample) as a Polynomial Chaos Expansion (PCE) surrogate $Y = ∑_{i∈ℕ} cᵢΨᵢ(𝐗)$ with:

$Ψᵢ$ being a polynomial basis function from polynomials, default gaussion/probabilists' Hermite of degree 8

$𝐗$ a random vector of distribution vector_distribution, default $𝐗∼𝒩(0,1)$,

$cᵢ$ the PCE coefficients to be calculated.

Functions

StatFEM_EUCLID.PCE.compute_statistics — Function

compute_statistics(pce::PolyChaosExpansion)

Compute mean and standard deviation of the given PCE

StatFEM_EUCLID.PCE.covariance — Function

covariance(pce::PolyChaosExpansion{T})::AbstractMatrix{T} where T <: Real

Compute covariance matrix of the given PCE

StatFEM_EUCLID.PCE.mean — Function

mean(pce::PolyChaosExpansion{T})::AbstractVector{T} where T<:Real

Compute mean values of the given PCE