Sampling

Module description

StatFEM_EUCLID.Sampling — Module

Sampling

This submodule provides all the high-level methods to sample the black box model through UM-Bridge.

Types

StatFEM_EUCLID.Sampling.UnivariateFEMSample — Type

struct UnivariateFEMSample{T<:Real}

Stores the result of a sampling call, i.e. the underlying uniform sample, the actual sample distribution and the obtained FEM sample

Functions

StatFEM_EUCLID.Sampling.compute_statistics — Method

compute_statistics(sample::UnivariateFEMSample{T}) where {T}

Compute empirical mean and standard deviation for the given sample

StatFEM_EUCLID.Sampling.sample_FEM — Method

sample_FEM(fem_model::HTTPModel, n_samples::Int;sample_distribution::UnivariateDistribution,rng = default_rng(),solution_index=1,config=empty_config())::UnivariateFEMSample

Sample the fem_model n_samples times with a univariate random parameter with distribution sample_distribution.

Optional keyword arguments:

rng: Specify a random number generator to use

solution_index: For a vector valued unknown, return the unknown at this index (can be : for the full solution)

config: Dict{String,Any} describing optional parameters for the fem_model